Hesitancy implemented
The hesitancy parameter is the probability each onset in color training will be skipped. This is intended to reduce overfitting that may have previously been caused by training on sequences back-to-back that were close in proximity (i.e. 1000 sequences from song A then 1000 sequences from song B, etc.). There are other ways to address this but the obvious techniques would require a lot of memory/storage.